test: Cleanup and extend conversion tests
